Extract bezier smoothing utility
